+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/ui/CCv2DSL.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -20,21 +20,40 @@ package com.intellij.idea.plugin.hybris.tools.ccv2.ui
 
 import com.intellij.ide.HelpTooltip
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
+import com.intellij.idea.plugin.hybris.notifications.Notifications
 import com.intellij.idea.plugin.hybris.settings.components.DeveloperSettingsComponent
 import com.intellij.idea.plugin.hybris.tools.ccv2.CCv2Util
 import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2EnvironmentDto
 import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2ServiceDto
+import com.intellij.notification.NotificationType
+import com.intellij.openapi.ide.CopyPasteManager
 import com.intellij.openapi.project.Project
 import com.intellij.ui.JBColor
 import com.intellij.ui.dsl.builder.Panel
 import com.intellij.ui.dsl.builder.RightGap
 import com.intellij.ui.dsl.builder.Row
+import java.awt.datatransfer.StringSelection
 import java.time.OffsetDateTime
 import javax.swing.Icon
 
 fun Row.date(label: String, dateTime: OffsetDateTime?) {
     label(CCv2Util.formatTime(dateTime))
         .comment(label)
+}
+
+fun Row.copyLink(project: Project, label: String?, value: String, confirmationMessage: String = "Copied to clipboard") {
+    link(value) {
+        CopyPasteManager.getInstance().setContents(StringSelection(value))
+        Notifications.create(NotificationType.INFORMATION, confirmationMessage, "")
+            .hideAfter(10)
+            .notify(project)
+    }
+        .comment(label)
+        .applyToComponent {
+            HelpTooltip()
+                .setTitle("Click to copy to clipboard")
+                .installOn(this);
+        }
 }
 
 fun Row.sUser(project: Project, sUserId: String, icon: Icon, label: String = "Created by") {